Add vtk ignores

+++ pyvista/core/filters/poly_data.py
@@ -4156,7 +4156,7 @@ class PolyDataFilters(DataSetFilters):
             _validation.check_instance(reference_volume, pyvista.ImageData, name='reference volume')
             # The image stencil filters do not support orientation, so we apply the
             # inverse direction matrix to "remove" orientation from the polydata
-            poly_ijk = self.transform(reference_volume.direction_matrix.T, inplace=False)
+            poly_ijk = self.transform(reference_volume.direction_matrix.T, inplace=False)  # type: ignore[misc]
             poly_ijk = _preprocess_polydata(poly_ijk)
         else:
             # Compute reference volume geometry
@@ -4262,9 +4262,9 @@ class PolyDataFilters(DataSetFilters):
             isinstance(val, (float, int)) and round(val) == val
             for val in (background_value, foreground_value)
         ):
-            scalars_dtype = int
+            scalars_dtype = np.int_
         else:
-            scalars_dtype = float
+            scalars_dtype = np.float64
         scalars = (  # Init with background value
             np.zeros(scalars_shape, dtype=scalars_dtype)
             if background_value == 0
@@ -4280,7 +4280,7 @@ class PolyDataFilters(DataSetFilters):
         poly_to_stencil = _vtk.vtkPolyDataToImageStencil()
         poly_to_stencil.SetInputData(poly_ijk)
         poly_to_stencil.SetOutputSpacing(*reference_volume.spacing)
-        poly_to_stencil.SetOutputOrigin(*reference_volume.origin)
+        poly_to_stencil.SetOutputOrigin(*reference_volume.origin)  # type: ignore[call-overload]
         poly_to_stencil.SetOutputWholeExtent(*reference_volume.extent)
         _update_alg(poly_to_stencil, progress_bar, 'Converting polydata')
